Provide single method for getting playwright browser versions
