+++ .github/workflows/release.yml
@@ -119,11 +119,11 @@ jobs:
         run: |
           node ./scripts/print-versions.js > ./browsers.json
 
-          CHROMIUM_VERSION=$(< browsers.json jq -r '.chromium')
+          CHROMIUM_VERSION=$(node ./scripts/print-browser-version.js 'chromium')
           echo "chromium_version=$CHROMIUM_VERSION" >> $GITHUB_ENV
-          FIREFOX_VERSION=$(< browsers.json jq -r '.firefox')
+          FIREFOX_VERSION=$(node ./scripts/print-browser-version.js 'firefox')
           echo "firefox_version=$FIREFOX_VERSION" >> $GITHUB_ENV
-          WEBKIT_VERSION=$(< browsers.json jq -r '.webkit')
+          WEBKIT_VERSION=$(node ./scripts/print-browser-version.js 'webkit')
           echo "webkit_version=$WEBKIT_VERSION" >> $GITHUB_ENV
 
       - name: Build and push
